+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 9.1.0
+
+- Added a support for the Instruction Manul
+- Requires Assorted Lib 4.1.0.
+
 ## 9.0.0
 
 Updated to Minecraft 26.2, for NeoForge and Fabric.